+++ tests/test_basin_mask.py
@@ -22,6 +22,7 @@ from hydromt.workflows.basin_mask import (
 logger = logging.getLogger("tets_basin")
 
 
+@pytest.mark.skip(reason="Needs refactor to Pydantic BaseModel.")
 def test_region(tmpdir, world, geodf, rioda):
     # prepare test data
     fn_gdf = str(tmpdir.join("world.geojson"))
@@ -136,6 +137,7 @@ def test_check_size(caplog):
     )
 
 
+@pytest.mark.skip(reason="Needs implementation of RasterDataSet.")
 @pytest.mark.filterwarnings("ignore::UserWarning")
 def test_basin(caplog):
     data_catalog = hydromt.DataCatalog("artifact_data", logger=logger)
